Moved getFiles to shared util
